+++ FactoryTest.php
@@ -62,6 +62,85 @@ final class FactoryTest extends TestCase
         ];
     }
 
+    /** @dataProvider provideValidData */
+    public function testFromData(string $data, string $mimetype, string $parameters, string $expected): void
+    {
+        self::assertSame($expected, Uri::fromData($data, $mimetype, $parameters)->toString());
+    }
+
+    public static function provideValidData(): iterable
+    {
+        yield 'empty data' => [
+            'data' => '',
+            'mimetype' => '',
+            'parameters' => '',
+            'expected' => 'data:text/plain;charset=us-ascii,',
+        ];
+
+        yield 'simple string' => [
+            'data' => 'Hello World!',
+            'mimetype' => '',
+            'parameters' => '',
+            'expected' => 'data:text/plain;charset=us-ascii,Hello%20World%21',
+        ];
+
+        yield 'changing the mimetype' => [
+            'data' => 'Hello World!',
+            'mimetype' => 'text/no-plain',
+            'parameters' => '',
+            'expected' => 'data:text/no-plain;charset=us-ascii,Hello%20World%21',
+        ];
+
+        yield 'empty mimetype but added parameters' => [
+            'data' => 'Hello World!',
+            'mimetype' => '',
+            'parameters' => 'foo=bar',
+            'expected' => 'data:text/plain;foo=bar,Hello%20World%21',
+        ];
+
+        yield 'empty data with optional argument field' => [
+            'data' => '',
+            'mimetype' => 'application/json',
+            'parameters' => 'foo=bar',
+            'expected' => 'data:application/json;foo=bar,',
+        ];
+
+        yield 'changing the parameters' => [
+            'data' => 'Hello World!',
+            'mimetype' => 'text/no-plain',
+            'parameters' => 'foo=bar',
+            'expected' => 'data:text/no-plain;foo=bar,Hello%20World%21',
+        ];
+
+        yield 'the parameters can start with the ; character' => [
+            'data' => 'Hello World!',
+            'mimetype' => 'text/no-plain',
+            'parameters' => ';foo=bar',
+            'expected' => 'data:text/no-plain;foo=bar,Hello%20World%21',
+        ];
+    }
+
+    public function testFromDataFailsWithInvalidMimeType(): void
+    {
+        $this->expectException(SyntaxError::class);
+
+        Uri::fromData('Hello World!', 'text\plain');
+    }
+
+    public function testFromDataFailsWithReservedParameterName(): void
+    {
+        $this->expectException(SyntaxError::class);
+
+        Uri::fromData('Hello World!', 'text/plain', 'base64=foobar');
+    }
+
+    public function testFromDataFailsWithMalformedParameters(): void
+    {
+        $this->expectException(SyntaxError::class);
+
+        Uri::fromData('Hello World!', 'text/plain', 'foobar');
+    }
+
     /**
      * @dataProvider unixpathProvider
      */