+++ Uri.php
@@ -507,6 +507,46 @@ final class Uri implements UriInterface
     }
 
     /**
+     * Create a new instance from a data string.
+     *
+     * @throws SyntaxError If the parameter syntax is invalid
+     */
+    public static function fromData(string $data, string $mimetype = '', string $parameters = ''): self
+    {
+        static $regexpMimetype = ',^\w+/[-.\w]+(?:\+[-.\w]+)?$,';
+
+        $mimetype = match (true) {
+            '' === $mimetype => 'text/plain',
+            1 === preg_match($regexpMimetype, $mimetype) =>  $mimetype,
+            default => throw new SyntaxError('Invalid mimeType, `'.$mimetype.'`.'),
+        };
+
+        if ('' != $parameters) {
+            if (str_starts_with($parameters, ';')) {
+                $parameters = substr($parameters, 1);
+            }
+
+            $validateParameter = function (string $parameter): bool {
+                $properties = explode('=', $parameter);
+
+                return 2 != count($properties) || 'base64' === strtolower($properties[0]);
+            };
+
+            $params = array_filter(explode(';', $parameters));
+            if ([] !== array_filter($params, $validateParameter(...))) {
+                throw new SyntaxError(sprintf('Invalid mediatype parameters, `%s`.', $parameters));
+            }
+
+            $parameters = ';'.$parameters;
+        }
+
+        return self::fromComponents([
+            'scheme' => 'data',
+            'path' => self::formatDataPath($mimetype.$parameters.','.rawurlencode($data)),
+        ]);
+    }
+
+    /**
      * Create a new instance from a Unix path string.
      */
     public static function fromUnixPath(Stringable|string $path): self
@@ -690,7 +730,7 @@ final class Uri implements UriInterface
     private function formatPath(string $path): string
     {
         return match (true) {
-            'data' === $this->scheme => Encoder::encodePath($this->formatDataPath($path)),
+            'data' === $this->scheme => Encoder::encodePath(self::formatDataPath($path)),
             'file' === $this->scheme => $this->formatFilePath(Encoder::encodePath($path)),
             default => Encoder::encodePath($path),
         };
@@ -703,7 +743,7 @@ final class Uri implements UriInterface
      *
      * @throws SyntaxError If the path is not compliant with RFC2397
      */
-    private function formatDataPath(string $path): string
+    private static function formatDataPath(string $path): string
     {
         if ('' == $path) {
             return 'text/plain;charset=us-ascii,';
@@ -726,7 +766,7 @@ final class Uri implements UriInterface
             $parameters = 'charset=us-ascii';
         }
 
-        $this->assertValidPath($mimetype, $parameters, $data);
+        self::assertValidPath($mimetype, $parameters, $data);
 
         return $mimetype.';'.$parameters.','.$data;
     }
@@ -738,7 +778,7 @@ final class Uri implements UriInterface
      *
      * @throws SyntaxError If the mediatype or the data are not compliant with the RFC2397
      */
-    private function assertValidPath(string $mimetype, string $parameters, string $data): void
+    private static function assertValidPath(string $mimetype, string $parameters, string $data): void
     {
         if (1 !== preg_match(self::REGEXP_MIMETYPE, $mimetype)) {
             throw new SyntaxError('The path mimetype `'.$mimetype.'` is invalid.');
@@ -749,7 +789,7 @@ final class Uri implements UriInterface
             $parameters = substr($parameters, 0, - strlen($matches[0]));
         }
 
-        $res = array_filter(array_filter(explode(';', $parameters), $this->validateParameter(...)));
+        $res = array_filter(array_filter(explode(';', $parameters), self::validateParameter(...)));
         if ([] !== $res) {
             throw new SyntaxError('The path paremeters `'.$parameters.'` is invalid.');
         }
@@ -767,7 +807,7 @@ final class Uri implements UriInterface
     /**
      * Validate mediatype parameter.
      */
-    private function validateParameter(string $parameter): bool
+    private static function validateParameter(string $parameter): bool
     {
         $properties = explode('=', $parameter);