+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All Notable changes to `League\Uri` will be documented in this file
 
+## Next - TBD
+
+### Added
+
+- None
+
+### Fixed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
 ## [7.3.0](https://github.com/thephpleague/uri/compare/7.2.1...7.3.0) - 2023-09-09
 
 ### Added